I have a few updates to add on the form I'll give you the JSON I tried: https://purprojet.slack.com/files/julian/F0SEQ78LE/questions_json.js you can try with this JSON for the questions there are several little bugs / options / or documentation to add: https://purprojet.slack.com/files/julian/F0SEXJW31/bug1.mov
TODO-1​ you can see the video, there is a small focus bug on some height/width ... probably only on the last question.... – can you fix it

TODO-2​ : the *LIST" choice is very good and very useful (`"type": "list", "multiple": true,`); however we sometimes need to add OTHERS choices which are not listed – so we should make a field "Other..." and when we click it, we can EDIT it, and then if it's MULTISELECT=1, another new "Other..." choice appears.
(so we can add/edit multiple non-initialy-listed choices)

​TODO-3​ when it goes to a next question, if new question is a "Text field", we need to FOCUS automatically on this (newInput.focus())
( because now, even if the new question is highlighed (opacity=1) but the problem is that OLD question INPUT is still focus if we type on keyboard)

​TODO-4​ sometimes we have 2 choices (Yes/No) and it would be good to VALIDATE as soon as we click YES/NO
Currently : Yes -> Enter -> go to next question or No -> Enter -> go to next question
I want : Yes -> go to next question ; No -> go to next question
You can add an option :
`quick-validate: TRUE/FALSE`
default is FALSE
if FALSE = same behavior as today,
if there is TRUE, then it goes to next qustion as soon as we click any choice.

​TODO-5​ currently it'es not possible to edit an old question, right?
can we make it possible? to edit/go back and edit and resubmit an old question?
we can add a global option for this :
`options.old_questions_editables = TRUE\FALSE` default FALSE like today